+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Whether the application uses the _graphql_, the _generateClientCode_ or the _gen
 * separateUtilityClasses: true _(both client and server mode)_
 * skipGenerationIfSchemaHasNotChanged: true _(both client and server mode)_
 
+# Not released yet
+
+Server mode:
+* Workaround to build a project with Gradle 7 without the gradle wrapper
 
 # 1.14.1